Revert "Combine demos (#14)"
This reverts commit fe8c2c8a2e12b179923b3546d1ebe46a1b086136.
