--- demo/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "jupyter-iframe-commands-demo",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "VITE_DEMO_SRC='./lite/index.html' VITE_TITLE='Lite' vite",
-    "build": "tsc && vite build --base=./",
-    "build:ghpages": "tsc && VITE_DEMO_SRC='./lite/index.html' VITE_TITLE='Lite' vite build --base=./",
-    "preview": "VITE_DEMO_SRC='./lite/index.html' VITE_TITLE='Lite' vite preview",
-    "start:lab": "jupyter lab --config jupyter_server_config.py",
-    "start:lite": "jlpm dev",
-    "start:local": "VITE_DEMO_SRC='http://localhost:8888' VITE_TITLE='Local' vite --port 8080"
-  },
-  "dependencies": {
-    "jupyter-iframe-commands-host": "0.1.0",
-    "sweetalert2": "^11.14.5"
-  },
-  "devDependencies": {
-    "typescript": "~5.6.2",
-    "vite": "^6.0.1"
-  }
-}
