+++ ghpages/index.html
@@ -1,0 +1,33 @@
+<!-- parent.html -->
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Demo</title>
+    <script type="module" src="src/main.js"></script>
+  </head>
+  <body>
+    <div
+      style="display: flex; align-items: center; justify-content: space-around"
+    >
+      <h1>Lite Demo</h1>
+      <div>
+        <form id="commands">
+          <input type="text" name="command" placeholder="Enter a command" />
+          <input type="text" name="args" placeholder="Enter args (optional)" />
+          <button type="submit">Submit</button>
+        </form>
+
+        <button id="list-commands">List Commands</button>
+      </div>
+    </div>
+
+    <iframe
+      id="jupyterlab"
+      src="./lite/index.html"
+      sandbox="allow-scripts allow-same-origin"
+      style="width: 100%; height: 100vh"
+    ></iframe>
+  </body>
+</html>
