+++ ghpages/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "jupyter-iframe-commands-demo",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "tsc && vite build --base=./",
+    "preview": "vite preview"
+  },
+  "dependencies": {
+    "jupyter-iframe-commands-host": "0.1.0"
+  },
+  "devDependencies": {
+    "typescript": "~5.6.2",
+    "vite": "^6.0.1"
+  }
+}
