+++ ghpages/src/style.css
@@ -1,0 +1,5 @@
+.demo-top {
+  display: flex;
+  align-items: center;
+  justify-content: space-around;
+}
