+++ ui-tests/package.json
@@ -5,12 +5,13 @@
   "private": true,
   "scripts": {
     "start:lab": "jupyter lab --config jupyter_server_test_config.py",
-    "start:server": "jlpm --cwd ../demo start:local",
+    "start:server": "jlpm web-dev-server -p 8080 --root-dir ../demo --node-resolve",
     "test": "jlpm playwright test",
     "test:update": "jlpm playwright test --update-snapshots"
   },
   "devDependencies": {
     "@jupyterlab/galata": "^5.0.5",
-    "@playwright/test": "^1.37.0"
+    "@playwright/test": "^1.37.0",
+    "@web/dev-server": "^0.4.6"
   }
 }
